[refactor] auth dip 준수 패키지 이동
